changed parser.hpp to have std::variant in vistor to parser
