+++ src/torro/include/scanner.hpp
@@ -81,112 +81,5 @@ class Scanner {
 
 
 
-class Visitor;
-class Expr{
-    public:
-    virtual std::string  accept(Visitor *) = 0;
-};
-class Binary : public Expr {
-    public:
-    Binary(std::shared_ptr<Expr> left, std::shared_ptr<Token> Operator, std::shared_ptr<Expr> right): left(left),Operator(Operator),right(right) {};    
-    std::shared_ptr<Expr> left;
-    std::shared_ptr<Token> Operator;
-    std::shared_ptr<Expr> right;
-    std::string accept(Visitor * );
-};
-class Grouping : public Expr {
-    public:
-    std::shared_ptr<Expr> expression;
-    Grouping(std::shared_ptr<Expr> expression):expression(expression){};
-    std::string accept(Visitor * );
-};
-
-class LiteralVisitor {
-  public:
-  std::string operator ()(std::string &v) {
-    return v;
-  }
-  std::string operator ()(std::monostate &v) {
-    return "void";
-  }
-  template <typename T>
-  std::string operator ()(T v) {
-    std::stringstream ss;
-    ss << v;
-    return ss.str();
-  }
-  
-  
-};
-
-class Literal : public Expr {
-    public:
-    typedef std::variant<std::monostate,bool,int,float,std::string> LiteralType;
-    LiteralType literal;
-    Literal (LiteralType literal): literal(literal) {};
-    Literal () {};
-    std::string accept(Visitor *);
-};
-
-class Unary : public Expr {
-    public:
-    std::shared_ptr<Token> Operator;
-    std::shared_ptr<Expr> expr;
-    Unary(std::shared_ptr<Token> Operator,std::shared_ptr<Expr> expr): Operator(Operator), expr(expr) {};
-    std::string accept(Visitor *);
-};
-
-
-class Visitor {
-    public:
-    virtual std::string visitBinary(std::shared_ptr<Binary> b) = 0 ;
-    virtual std::string visitGrouping(std::shared_ptr<Grouping> b) = 0 ;
-    virtual std::string visitLiteral(std::shared_ptr<Literal> b) = 0 ;
-    virtual std::string visitUnary(std::shared_ptr<Unary> b) = 0 ;
-    
-};
-class AstPrinter : Visitor { 
-    
-    template <class Head,class ...Args>
-    std::string parenthesize(Head head,Args ...args) {
-        std::stringstream  ss;
-        ss << "(" << handleCase(head) << " ";
-        ss << parenthesize(args...);
-        ss << ")";
-        return ss.str();                             
-    }                   
-    template <class T>
-    std::string parenthesize(T t) {
-        std::stringstream  ss;
-        ss << handleCase(t);
-        return ss.str();                             
-    }                   
-
-    std::string handleCase(std::shared_ptr<Expr> &e) {
-        return e->accept(this);
-    };
-    std::string handleCase(std::string s) {
-        return s;
-    };
-    public:
-    std::string print(std::shared_ptr<Expr> &expr) {                                            
-        return expr->accept(this);                                          
-    }
-    std::string visitBinary(std::shared_ptr<Binary> binary) override {
-        return parenthesize(binary->Operator->lexeme, binary->left, binary->right);
-    }
-    std::string visitGrouping(std::shared_ptr<Grouping> group) override{
-        return parenthesize("group",group->expression);  
-    }
-    std::string visitLiteral(std::shared_ptr<Literal> literal) override {                
-      if (!std::holds_alternative<std::monostate>(literal->literal)) 
-        return "nil";
-      return std::visit(LiteralVisitor{},literal->literal);                            
-    }   
-    std::string visitUnary(std::shared_ptr<Unary> unary) override {                    
-        return parenthesize(unary->Operator->lexeme, unary->expr);           
-  }     
-
-};
 
 #endif
